backend/requirements.txt
# FastAPI勤怠管理システム - 依存関係
# 生成日: 2025-12-29

# === Core Framework ===
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# === Database ===
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.13.1

# === Authentication & Security ===
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-jwt==4.0.0

# === Environment & Configuration ===
pydantic==2.5.1
pydantic-settings==2.1.0
python-dotenv==1.0.0

# === HTTP Client ===
httpx==0.25.2
aiofiles==23.2.1

# === CORS ===
python-cors==1.7.0

# === Validation & Utilities ===
email-validator==2.1.0
phonenumbers==8.13.26

# === Development & Testing ===
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-httpx==0.26.0

# === Production ===
gunicorn==21.2.0